requirements.txt
pyserial==3.4
tqdm==4.48.2
